Move keccak2x into own folder to avoid conflicts
